move save_fixture() to new tests/_helpers.py along with stuff in tests/__init__.py
