+++ tests/test_parity.py
@@ -7,7 +7,7 @@ from ml_matrics import (
     scatter_with_err_bar,
 )
 
-from . import y_pred, y_true
+from ._helpers import y_pred, y_true
 
 
 def test_density_scatter():
