+++ tests/test_utils.py
@@ -3,7 +3,7 @@ from matplotlib.offsetbox import AnchoredText
 
 from ml_matrics.utils import add_mae_r2_box, get_crystal_sys
 
-from . import y_pred, y_true
+from ._helpers import y_pred, y_true
 
 
 def test_add_mae_r2_box():
